+++ src/sksl/codegen/SkSLSPIRVCodeGenerator.h
@@ -24,7 +24,6 @@
 #include "src/sksl/ir/SkSLVariable.h"
 #include "src/sksl/spirv.h"
 
-#include <cstddef>
 #include <cstdint>
 #include <memory>
 #include <string_view>
@@ -446,8 +445,8 @@ private:
     SpvId toComponent(SpvId id, int component);
 
     struct ConditionalOpCounts {
-        size_t numReachableOps;
-        size_t numStoreOps;
+        int numReachableOps;
+        int numStoreOps;
     };
     ConditionalOpCounts getConditionalOpCounts();
     void pruneConditionalOps(ConditionalOpCounts ops);
